Add files via upload
Databases Coursework:
In this coursework I was tasked with creating a database system for the smooth and efficient operation of a business of my choosing. I specified what type of business I will be building my databases system for, along with the different collections of data that may be required to be stored in my database in
order to achieve its purpose. The completion of my coursework was graded by submission of a written report as well as a script of SQL statements.